move deployment class to flowsheet manager file; frontend provided project to determine deployment, app settings; add api endpoint for setting project
